AIM-181 vlm model code files refactoring.
