+++ .gitignore
@@ -18,6 +18,7 @@ vlm_model/constants/__pycache__/
 vlm_model/routers/__pycache__/
 vlm_model/schemas/__pycache__/
 vlm_model/utils/__pycache__/
+vlm_model/utils/analysis_video/__pycache__/
 
 # Ignore all log files within logs/ directory
 logs/*.log
