+++ vlm_model/utils/analysis.py
@@ -19,77 +19,6 @@ logger = logging.getLogger(__name__)
 # OpenAI 모듈을 client로 정의
 client = OpenAI() # openai
 
-def load_user_prompt() -> str:
-    """
-    프롬프트 파일을 로드합니다.
-    """
-    # 현재 파일의 위치를 기준으로 상대 경로 설정
-    current_dir = Path(__file__).parent
-    prompt_path = current_dir.parent.parent / 'prompt.txt'  # 실제 프로젝트 구조에 맞게 조정
-    
-    try:
-        with prompt_path.open('r', encoding='utf-8') as file:
-            return file.read()
-    except FileNotFoundError as e:
-        logger.error(f"프롬프트 파일을 찾을 수 없음: {prompt_path}", extra={
-            "errorType": "FileNotFoundError",
-            "error_message": f"프롬프트 파일을 찾을 수 없음: {prompt_path}"
-        })
-        raise PromptImportingError("프롬프트 파일을 찾을 수 없습니다") from e
-    except Exception as e:
-        logger.error(f"프롬프트 파일을 로드할 수 없음: {e}", extra={
-            "errorType": type(e).__name__,
-            "error_message": str(e)
-        })
-        raise PromptImportingError("프롬프트 파일을 로드하는 중 오류가 발생했습니다") from e
-
-def parse_feedback_text(feedback_text: str) -> FeedbackSections:
-    """
-    feedback_text를 FeedbackSections 형식으로 파싱합니다.
-    """
-    try:
-        feedback_json = json.loads(feedback_text)
-        if feedback_json.get("problem") == "none":
-            return FeedbackSections(
-                gaze_processing=FeedbackDetails(improvement="", recommendations=""),
-                facial_expression=FeedbackDetails(improvement="", recommendations=""),
-                gestures=FeedbackDetails(improvement="", recommendations=""),
-                posture_body=FeedbackDetails(improvement="", recommendations=""),
-                movement=FeedbackDetails(improvement="", recommendations="")
-            )
-        
-        # 섹션별로 데이터 추출
-        feedback_data = {}
-        for section_key, field_name in {
-            "gaze_processing": "gaze_processing",
-            "facial_expression": "facial_expression",
-            "gestures": "gestures",
-            "posture_body": "posture_body",
-            "movement": "movement"
-        }.items():
-            section = feedback_json.get(section_key, {})
-            improvement = section.get("improvement", "").strip()
-            recommendations = section.get("recommendations", "").strip()
-            feedback_data[field_name] = FeedbackDetails(
-                improvement=improvement,
-                recommendations=recommendations
-            )
-        
-        return FeedbackSections(**feedback_data)
-    
-    except json.JSONDecodeError as e:
-        logger.error(f"JSON 디코딩 오류: {str(e)}", extra={
-            "errorType": "JSONDecodeError",
-            "error_message": str(e)
-        })
-        raise HTTPException(status_code=400, detail="JSON 디코딩 오류") from e
-    except Exception as e:
-        logger.error(f"FeedbackSections 생성 실패: {str(e)}", extra={
-            "errorType": type(e).__name__,
-            "error_message": str(e)
-        })
-        raise HTTPException(status_code=500, detail="FeedbackSections 생성 실패") from e
-
 def analyze_frames(frames: List[np.ndarray], segment_idx: int, duration: int, segment_length: int, system_instruction: str, frame_interval: int = 3) -> Tuple[List[Tuple[np.ndarray, int, int, str]], List[str]]:
     """
     주어진 프레임들을 분석하여 문제 행동을 감지하고 피드백을 생성합니다.
